pyproject.toml
[project]
name = "dem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=0.3.2",
    "mcp-use>=1.3.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
